tests/directoryentry-manual.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Entries API: DirectoryEntry manual test</title>
<link rel=help href="https://wicg.github.io/entries-api/#api-directoryentry">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/idlharness.js"></script>

<h1>Entries API: DirectoryEntry manual test</h1>

<div id=drop style="height: 50px; border: 1px dotted red;">Drop the <b>support/upload</b> directory here.</div>
<script src="support.js"></script>

<script>
entry_test((t, entry) => {
  // Entry ----------

  assert_idl_attribute(entry, 'isFile', 'DirectoryEntry has isFile attribute');
  assert_equals(typeof entry.isFile, 'boolean', 'isFile is boolean');

  assert_idl_attribute(entry, 'isDirectory', 'DirectoryEntry has isDirectory attribute');
  assert_equals(typeof entry.isDirectory, 'boolean', 'isDirectory is boolean');

  assert_idl_attribute(entry, 'name', 'DirectoryEntry has name attribute');
  assert_equals(typeof entry.name, 'string', 'name is a string');

  assert_idl_attribute(entry, 'fullPath', 'DirectoryEntry has fullPath attribute');
  assert_equals(typeof entry.fullPath, 'string', 'fullPath is a string');

  assert_idl_attribute(entry, 'filesystem', 'DirectoryEntry has filesystem attribute');
  assert_class_string(entry.filesystem, 'DOMFileSystem', 'filesystem is a DOMFileSystem');

  assert_idl_attribute(entry, 'getParent', 'DirectoryEntry has getParent');
  assert_equals(typeof entry.getParent, 'function', 'getParent() is a method');

  // DirectoryEntry  ----------

  assert_false('DirectoryEntry' in self, 'Entry has [NoInterfaceObject]');
  assert_class_string(entry, 'DirectoryEntry', 'interface name is DirectoryEntry');

  assert_idl_attribute(entry, 'getFile', 'DirectoryEntry has getFile');
  assert_equals(typeof entry.getFile, 'function',
                'getFile() is a method');

  assert_idl_attribute(entry, 'getDirectory', 'DirectoryEntry has getDirectory');
  assert_equals(typeof entry.getDirectory, 'function',
                'getDirectory() is a method');

  assert_idl_attribute(entry, 'createReader', 'DirectoryEntry has createReader');
  assert_equals(typeof entry.createReader, 'function',
                'createReader() is a method');

  t.done();
}, 'DirectoryEntry - IDL');

entry_test((t, entry) => {
  assert_false(entry.isFile, 'isFile is false');
  assert_true(entry.isDirectory, 'isDirectory is true');
  assert_equals(entry.name, 'upload', 'expected directory was uploaded');
  assert_equals(entry.fullPath, '/upload', 'directory is child of root directory');
  t.done();
}, 'DirectoryEntry - attributes');
</script>
